+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## 2021-12-13
 ### Changed
 - AdvancedFor now allows multiple variable imports. Imports will be included at cursor position.
+- Increased minimum width of AdvancedFor